backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolioai-backend"
version = "0.1.0"
description = "Backend for PortfolioAI application"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "sqlalchemy>=1.4.0",
    "pydantic>=1.8.0",
    "python-dotenv>=0.19.0",
    "httpx>=0.19.0",
    "jinja2>=3.0.0",
    "python-docx>=0.8.11",
    "PyPDF2>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0",
    "isort>=5.0.0",
    "mypy>=0.900"
]

[tool.setuptools]
package-dir = { "" = "." }
